Fixes double-join glitch.
Previously, the "Cancel" button on the game queue UI in the main menu would
fail to work after the first game in a session. This has been fixed.

Further, the server has been patched to ignore join queue requests from
sessions which are already in the game queue, fixing this bug on both the
client and the server.