Tutorials are implemented.
Also fixed some assets and improved font contrast. Will probably need to tweak the font slightly.

Also, of note, it's hard for the follower to tell what's up with the card they
just stepped on, as they can't see the card when they're on top of it. Maybe
add an improved card collection graphic?