Major DB migration + visualization script addons.
This change adds an "orientation_before" column to the Move table. For moves
where this value did not exist, the default is 0. Which is wrong.
db_tools/integrate_orientation is a one-time tool to comb the database and
replay games to calculate the orientation_before value for all moves.

Oops, glad I caught this now, but comforting to know that in around an hour or
so I was able to recover the missing data because we store everything with
enough redundancy to make up for it.

Oh and visualization now draws the initial player orientation (motivating change
that caught this messup).