Server-side mitigation to invisible cards glitch.
This is an attempt to further-prevent the invisible cards glitch.

While this was already fixed in 64f3cd1, this change makes it so that even if a
similar map update glitch occurs, it will be mitigated. Map updates are sent
every time the role switches (Leader/Follower). This means that stale or
incorrectly applied initial map updates will be replaced.