Fixed keyboard shortcut feedback spam issue.
Previously, using live-feedback keyboard shortcuts in the client would spam the
server with thousands of messages (roughly 200kb/s). The logic has been changed
to only send on key-down instead of while the key is held.

Additionally, we discovered that with the keyboard shortcuts, the client would
continue to send live feedback, even out-of-turn (or from the follower's
client!). This shockingly wasn't verified on the server, and instead committed
to the database. This issue is also fixed in this update.