PyClient API improvement.
Step() previously returned a large tuple. This isn't great as users of the
function must figure out which element goes where, and when reading example code
which uses the API, it's not always obvious what elements are what (for example,
unused values are unpacked as "_", which can be confusing). Instead, we now
create a GameState object which is compatible with old use cases. Since the
GameState object is iterable, it can be unpacked -- this means no code changes
are required by clients. However it does enable a better experience in the API,
as the return value now has type hinting. Also, future code can be made more
readable.... e.g.:

Instead of:

```
map, _, _, instructions, _, _ = game.step(action)
```

We have

```
game_state = game.step(action)
```