Demo mode long-term stability improvment.
Demo mode no longer displays "Game Over" when transitioning from the end of one
game to the start of the next.

Also, action duration expiration is fixed. This is actually going backwards...
we previously used utc time for expiration, and this commit reverts it back to
local time due to replay mode in the client not being compiled for that. Will
need to push a fix to the client later.