Makes it so that card covers are configurable per-lobby.
Also makes it so that border coloring and card selection sounds are adequately
controlled by the card cover setting -- if card covers are enabled, the follower
only sees card borders as blue, even if invalid (and the contrapositive). This
is also true for sounds that would indicate whether or not the card selected
created an invalid set.

Also adds a rate limit for upload of client bugs, to avoid overwhelming the
server and also to avoid performance degradation of client.